[17] Restructuring server and adding mongo support.
